+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # tailwind-factory
 
+## 1.1.1
+
+### Patch Changes
+
+- removing en-us readme
+
+## 1.1.0
+
+### Minor Changes
+
+- support for boolean values in variants
+
+## 1.0.5
+
+### Patch Changes
+
+- removing console log and improving default variants type
+
 ## 1.0.4
 
 ### Patch Changes